+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Change log
 
-### 9.1.6-b0 (2/10/2025 - )
+### 9.1.6 (2/10/2025 - 2/25/2025)
 
 * http_client: tweak sse checking
 * undertow: updated to 2.3.18.Final
@@ -8,7 +8,8 @@
 * sse: support method PUT/POST with body
 * sse: channel.close() now closes sse connection gracefully
 * sse: change retry instruction to 5s
-  > in cloud env, we set backend timeout to around 620s, and let client (js) reconnect every 600s
+  > in cloud env, backend timeout is set to 600s
+* sse: channel.send() returns if sent successfully
 
 ### 9.1.5 (11/11/2024 - 01/22/2025)